Move the few remains of ui.py to launcher.py a
