+++ cddagl/launcher.py
@@ -1,24 +1,27 @@
-import sys
-import os
-import traceback
-
 import logging
+import os
+import sys
+import traceback
+from io import StringIO
 from logging.handlers import RotatingFileHandler
 
-from io import StringIO
-
+from PyQt5.QtGui import QIcon
+from PyQt5.QtWidgets import QApplication
 from babel.core import Locale
 
+import cddagl
 import cddagl.constants as cons
-from cddagl.constants import get_cddagl_path, get_locale_path
+from cddagl.constants import get_cddagl_path
+from cddagl.constants import get_locale_path, get_resource_path
+from cddagl.i18n import load_gettext_locale
 from cddagl.i18n import load_gettext_no_locale, proxy_gettext as _, get_available_locales
 from cddagl.sql.functions import init_config, get_config_value, config_true
-from cddagl.ui_old import start_ui, ui_exception
-
+from cddagl.ui.views.dialogs import ExceptionWindow
+from cddagl.ui.views.tabbed import TabbedWindow
 from cddagl.win32 import get_ui_locale, SingleInstance, write_named_pipe
 
-import cddagl
 version = cddagl.__version__
+logger = logging.getLogger('cddagl')
 
 
 def init_single_instance():
@@ -119,6 +122,39 @@ def handle_exception(extype, value, tb):
         .format(version=version, extype=str(extype), value=str(value),traceback=tb_io.getvalue())
     )
     ui_exception(extype, value, tb)
+
+
+def start_ui(locale, single_instance):
+    load_gettext_locale(get_locale_path(), locale)
+
+    main_app = QApplication(sys.argv)
+    main_app.setWindowIcon(QIcon(get_resource_path('launcher.ico')))
+
+    main_win = TabbedWindow('CDDA Game Launcher')
+    main_win.show()
+
+    main_app.main_win = main_win
+    main_app.single_instance = single_instance
+    main_app.app_locale = locale
+    sys.exit(main_app.exec_())
+
+
+def ui_exception(extype, value, tb):
+    main_app = QApplication.instance()
+
+    if main_app is not None:
+        main_app_still_up = True
+        main_app.closeAllWindows()
+    else:
+        main_app_still_up = False
+        main_app = QApplication(sys.argv)
+
+    ex_win = ExceptionWindow(main_app, extype, value, tb)
+    ex_win.show()
+    main_app.ex_win = ex_win
+
+    if not main_app_still_up:
+        sys.exit(main_app.exec_())
 
 
 def init_exception_catcher():
